Generalize ray caster and move it to the XrScene so that other scenes can override the objectInteraction method and implement their own interactions using the ray caster.
